Ya casi acabado

+++ software/models/VentaDetalleModel.py
@@ -7,8 +7,8 @@ class VentaDetalle(models.Model):
     idventadetalle = models.AutoField(primary_key=True)
     idventa = models.ForeignKey(Venta, models.DO_NOTHING, db_column='idventa')
     idproducto = models.ForeignKey(Producto, models.DO_NOTHING, db_column='idproducto')
-    cantidad = models.IntegerField()
-    preciosubtotal = models.FloatField()
+    cantidad = models.DecimalField(max_digits=10, decimal_places=2)
+    preciosubtotal = models.DecimalField(max_digits=10, decimal_places=2)
 
     class Meta:
         managed = False
